CB_Sandbox/requirements-lock.txt
alembic==1.16.4
altair==4.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
boto3==1.40.7
botocore==1.40.7
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
databricks-sdk==0.62.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
docker==7.1.0
dynaconf==3.2.11
entrypoints==0.4
evidently==0.7.12
Faker==37.5.3
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
great_expectations==1.5.8
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lightgbm==4.6.0
litestar==2.17.0
litestar-htmx==0.5.0
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
mistune==3.1.3
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
mock==4.0.3
msgspec==0.19.0
multidict==6.6.4
multipart==1.3.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==24.2
pandas==2.1.4
pathos==0.3.4
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
polyfactory==2.22.1
posthog==6.5.0
pox==0.3.6
ppft==1.7.7
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rich-click==1.8.9
rpds-py==0.27.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
sagemaker==2.250.0
sagemaker-core==1.0.50
schema==0.7.7
scikit-learn==1.7.1
scipy==1.16.1
shap==0.48.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
statsmodels==0.14.5
tblib==3.1.0
tenacity==9.1.2
threadpoolctl==3.6.0
toolz==1.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xgboost==3.0.4
zipp==3.23.0
